Rename docs test to dec test and expand DEC coverage
